+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 ## Release 1.0.3
 
 #### Migrations
- - GithubRunner has been updated to work with the new workflows included in examples. This requires use of the schema_map.json file and having the workflows use the schema names, rather than JSON encoded schemas. This ensures that any schema run by the bot is present in the repo.
+ - GithubRunner has been updated to work with the new workflows included in examples. This requires use of the schema_map.json file and having the workflows use the schema names, rather than JSON encoded schemas. This ensures that any schema run by the bot is present in the repo. Replace existing workflows with new workflows in `examples/workflows`.
+ - Updated migration commands.
+  - atmigrate-1.0.1 to migrate Manager for 1.0.1 changes.
+  - atmigrate-1.0.3 to migrate Scheduler/Schema Map for 1.0.3 changes.
 
 #### Documentation
  - Added documentation for creating and using custom components, including an example PR.